Move the bare-prefix unneeded-suppression regression test to jtreg
checker/jtreg/nullness/UnneededSuppressionsTest.java already covers
-AwarnUnneededSuppressions with NullnessChecker for message-key-qualified
suppressions; add the bare-prefix case (a suppression whose value is
exactly a checker prefix, suppressing every warning of that checker,
including the unneeded.suppression warning about itself) as three more
methods there instead of a separate JUnit test and test-input directory.

Confirmed the new methods fail the same way the old JUnit test did
against the pre-fix SourceChecker.java (checked out the parent commit,
reran, restored the fix) and pass with it; the full jtreg suite remains
green.
